ofproto-dpif-upcall: Remove redundant 'packets' list from struct flow_miss.
Until now, struct flow_miss contained a list of packets and a list of
upcalls.  Each packet in the list of packets can be obtained from the
corresponding upcall in the list of upcalls via upcall->dpif_upcall.packet,
so this commit deletes the list of packets and replaces each reference to
a packet by that expression.

Signed-off-by: Ben Pfaff <blp@nicira.com>
Acked-by: Jarno Rajahalme <jrajahalme@nicira.com>