datapath: Check for backported removal of HAVE_NET_DEVICE_OPS
The symbol HAVE_NET_DEVICE_OPS was removed in 3.1 even though these ops
still exist, so on newer kernels we define the symbol ourselves. However,
SLES 11 SP3 appears to have backported this symbol removal. This simply
defines the symbol on all kernels where it is relevant, unless it is
already defined.

Reported-by: Cristina Palomo Regidor <cristina.palomo.regidor@ericsson.com>
Signed-off-by: Jesse Gross <jesse@nicira.com>
Acked-by: Pravin B Shelar <pshelar@nicira.com>