tests: Remove the unit test for "ovs-appctl coverage/show".
Changes will be made to allow "ovs-appctl time/warp" wake up all
threads.  With that, all threads will try calling "coverage_run()".
And it will be impossible to deterministically check the output of
the "ovs-appctl coverage/show" command.  Thus, this commit removes
the unit test for the coverage/show command.

Signed-off-by: Alex Wang <alexw@nicira.com>
Signed-off-by: Ethan Jackson <ethan@nicira.com>
Acked-by: Ethan Jackson <ethan@nicira.com>